app/requirements.txt
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.11.0
altgraph==0.17
appdirs==1.4.4
asn1crypto==1.4.0
bcrypt==3.1.7
build==1.2.2.post1
cachecontrol==0.14.2
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==3.4.3
cleo==2.1.0
click==7.1.2
colorama==0.4.3
cos-python-sdk-v5==1.7.7
crashtest==0.4.1
crcmod==1.7
cryptography==45.0.7
Cython==0.29.21
dicttoxml==1.7.4
distlib==0.4.0
dnspython==2.0.0
dulwich==0.21.7
fastjsonschema==2.21.2
filelock==3.16.1
Flask==2.0.3
Flask-Cors==5.0.0
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-core==1.4.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
identify==1.4.25
idna==2.10
importlib-metadata==8.5.0
importlib-resources==3.0.0
installer==0.7.0
IPy==1.0
itsdangerous==2.1.1
jaraco.classes==3.4.0
Jinja2==3.0.3
jmespath==0.10.0
jwt==1.3.1
keyring==24.3.1
lxml==4.5.2
MarkupSafe==2.1.0
more-itertools==10.5.0
msgpack==1.1.1
mysql-connector-python==9.0.0
mysqlclient==2.0.1
Naked==0.1.31
ndg-httpsclient==0.5.1
nodeenv==1.4.0
oss2==2.5.0
packaging==25.0
paramiko==2.7.1
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.9.0
Pillow==7.2.0
pkginfo==1.12.1.2
platformdirs==4.3.6
pluggy==0.13.1
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pre-commit==2.6.0
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.7.0
py==1.9.0
py7zr==0.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
PyInstaller==3.6
pymemcache==3.2.0
pymongo==3.11.0
pymssql @ file:///C:/pymssql-2.1.4-cp38-cp38-win_amd64.whl
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyotp==2.4.0
pyparsing==2.4.7
pyproject-hooks==1.2.0
python-dateutil==2.8.2
python-engineio==3.13.1
python-memcached==1.59
python-socketio==4.6.0
python3-memcached==1.51
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qiniu==7.2.8
rapidfuzz==3.9.7
redis==3.5.3
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.6
shellescape==3.8.1
shellingham==1.5.4
six==1.15.0
some-package==0.1
SQLAlchemy==1.3.18
tencentcloud-sdk-python==3.0.532
texttable==1.6.2
toml==0.10.1
tomli==2.2.1
tomlkit==0.13.3
tox==3.18.1
translate==3.5.0
trove-classifiers==2025.9.11.17
typing-extensions==4.13.2
unrar==0.4
upyun==2.5.5
urllib3==1.25.10
virtualenv==20.34.0
watchdog==0.10.3
websocket-client==1.2.1
Werkzeug==2.0.3
winshell==0.6
WMI==1.5.1
xmltodict==0.12.0
zipp==3.20.2
zope.event==4.4
zope.interface==5.1.0
